updating Logan-classes

+++ Embedded-Proj-B/Core/Inc/inputDriver.h
@@ -22,7 +22,7 @@ public:
     int8_t pinNumB;
     //Functions
     inputDriver(GPIO_TypeDef* , GPIO_TypeDef* ,uint8_t,uint8_t);
-    uint32_t edgeDetector();
+    int8_t edgeDetector();
 };
 
 #endif /* INPUTDRIVER_H_ */
